Fixed typo. Added new line after exception messages. Added try/catch to complete example.

+++ examples/complete_example.cpp
@@ -10,41 +10,56 @@ int main()
 {
 	/*** LOADING FILES  *************************************************************************************/
 
-	// Loading an XML file is easy, and it returns a pointer to the root node.
-	xml::Node* root = xml::loadXml("example.xml");
 
-	/*** PRINTING THE STRUCTURE *****************************************************************************/
+	xml::Node* root = NULL;
 
-	// You can print the XML structure in a readable format.
-	std::cout << "Printing the XML structure..." << std::endl;
-	xml::printTree(root);
-	std::cout << std::endl;
+	// Malformed XML can cause an exception to be thrown when loadXml is called.
+	try
+	{
+		// Loading an XML file is easy, and it returns a pointer to the root node.
+		root = xml::loadXml("example.xml");
+	}
+	catch (const xml::EasyXmlException& e)
+	{
+		// Print out the error message.
+		std::cout << e.what();
+	}
 
-	/*** ACCESSING NODES ************************************************************************************/
+	if (root != NULL)
+	{
+		/*** PRINTING THE STRUCTURE *****************************************************************************/
 
-	// You can search for nodes by name, using "/" as a hierarchical delimiter.
-	// findNode() will throw an exception if the given node was not found OR you pass the boolean "true"
-	// as the second argument and it will return NULL instead.
-	xml::Node* n2008 = root->findNode("history/2008");
+		// You can print the XML structure in a readable format.
+		std::cout << "Printing the XML structure..." << std::endl;
+		xml::printTree(root);
+		std::cout << std::endl;
 
-	// You can access a node's name and value directly so you can read or make changes to the XML structure
-	std::cout << "The value of \"" << n2008->name << "\" is \"" << n2008->value << "\""
-	          << std::endl;
+		/*** ACCESSING NODES ************************************************************************************/
 
-	xml::Node* year = root->findNode("year", true);
+		// You can search for nodes by name, using "/" as a hierarchical delimiter.
+		// findNode() will throw an exception if the given node was not found OR you pass the boolean "true"
+		// as the second argument and it will return NULL instead.
+		xml::Node* n2008 = root->findNode("history/2008");
 
-	// There is also a templated "getter" method for value that allows you to choose between std::string, int,
-	// float, or double as the return type.
-	std::cout << "The integer value of \"" + year->name + "\" + 1 is " << year->val<int>() + 1 \
-	          << std::endl;
+		// You can access a node's name and value directly so you can read or make changes to the XML structure
+		std::cout << "The value of \"" << n2008->name << "\" is \"" << n2008->value << "\""
+		          << std::endl;
 
-	xml::saveXml(root, "example_output.xml");
+		xml::Node* year = root->findNode("year", true);
 
-	/*** CLEANUP ********************************************************************************************/
-	          
-	// When finished with an XML tree, don't forget to free the memory.
-	// deleteTree will recursively delete a node and its children for you.
-	xml::deleteTree(root);
+		// There is also a templated "getter" method for value that allows you to choose between std::string, int,
+		// float, or double as the return type.
+		std::cout << "The integer value of \"" + year->name + "\" + 1 is " << year->val<int>() + 1 \
+		          << std::endl;
+
+		xml::saveXml(root, "example_output.xml");
+
+		/*** CLEANUP ********************************************************************************************/
+		          
+		// When finished with an XML tree, don't forget to free the memory.
+		// deleteTree will recursively delete a node and its children for you.
+		xml::deleteTree(root);
+	}
 
 	return 0;
 }
